#19 refactor: 완료 상태의 변수명 변경
